fix: Add registerNetworks in createTokenBridgePrepareSetWethGatewayTransactionRequest
